feat(drivemotor): make ticks_per_meter fully runtime configurable
- add ticks_per_meter to SET_DRIVE_PID v2 protocol
- propagate runtime ticks_per_meter from GUI to STM32
- replace compile-time TICKS_PER_M usage with runtime value
- preserve wheel_tick_factor precision as float end-to-end
- update wheel odometry, PI controller and glitch filtering
- add protocol compatibility guards and validation
- add protocol, persistence and report regression tests

+++ tools/motor/mowgli_tools/drive_pid_tuner.py
@@ -253,7 +253,7 @@ class DrivePidTuner(Node):
         self._latest_gnss_status: GnssStatus | None = None
         self._latest_odom_time: float | None = None
         self._latest_wheel_tick_time: float | None = None
-        self._latest_wheel_tick_factor: int | None = None
+        self._latest_wheel_tick_factor: float | None = None
         self._latest_wheel_tick_stamp: str | None = None
         self._failure_message: str | None = None
         self._failure_status_snapshot: dict[str, Any] | None = None
@@ -319,7 +319,7 @@ class DrivePidTuner(Node):
 
     def _on_wheel_ticks(self, msg: WheelTick) -> None:
         self._latest_wheel_tick_time = time.monotonic()
-        self._latest_wheel_tick_factor = int(msg.wheel_tick_factor)
+        self._latest_wheel_tick_factor = float(msg.wheel_tick_factor)
         self._latest_wheel_tick_stamp = self._stamp_to_iso(msg.stamp)
         if self._active_trial is None:
             return
